+++ ci4/app/Interfaces/DeploymentPodMetrics.php
@@ -1,0 +1,10 @@
+<?php namespace App\Interfaces;
+/**
+ * @property string $pod
+ * @property string $container
+ * @property int $cpu_millicores
+ * @property int $memory_bytes
+ */
+interface DeploymentPodMetrics {
+
+}